Made base class for the copy methods
